Fix two Int64 overflows, accept any AbstractVector, and right-size the interval warnings
Three fixes, none changing any value a test already pinned:

- The signed rank variance formed nz(nz+1)(2nz+1) in Int64, which wraps
  from nz = 2^21: at exactly 2^21 the product wraps to a positive wrong
  value, so sigma came out three orders of magnitude too small with no
  error raised. tiedrank_adj had the same problem in t^3 for a tie group
  of 2^21 or more equal values. Both products now form in Int128; below
  the old wrap point the results are bit-identical.

- The signed rank constructors declared AbstractVector but passed the
  input straight to struct fields typed Vector, so ranges and views threw
  a MethodError. They convert on the way in now, as the Mann-Whitney side
  always has. (Pre-existing upstream, not introduced by this stack.)

- Two warning fixes in the interval index rules. exact_ci_index warned
  "level not attainable" when cdf(0) equalled alpha/2 exactly, a case
  where the k = 0 interval attains the request; the comparison is now
  strict. And for one-sided intervals both rules reported the two-sided
  2*level - 1 as the request and a two-sided attainable coverage; they
  now report the level the caller asked for and the one-sided bound
  1 - cdf(0). The upper clamp also tightens to div(m - 1, 2), whose
  even-m endpoint pair cannot invert (the old div(m, 2) was unreachable
  anyway, by P(W <= m/2) >= 1/2 > alpha/2).

+++ src/common.jl
@@ -45,7 +45,8 @@ function tiedrank_adj!(ord::AbstractVector, v::AbstractArray)
         if j > i
             t = j - i + 1
             m = sum(i:j) / t
-            tieadj += t^3 - t
+            # in Int128: t^3 wraps Int64 for a tie group of 2^21 or more equal values
+            tieadj += Int128(t)^3 - t
             for k = i:j
                 ord[place[k]] = m
             end
